pyproject.toml
[project]
name = "idn-area-etl"
version = "0.1.0"
description = "CLI for extracting indonesia area codes and names from PDF tables and converting them to structured CSV format."
authors = [
    { name = "fityannugroho", email = "fityannugroho@gmail.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py>=1.0.9",
    "ghostscript>=0.8.1",
    "pandas>=2.0.3",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.2.2.240807",
    "pyright>=1.1.405",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.12.8",
]

[project.scripts]
idnareaetl = "idn_area_etl.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.uv]
package = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: end-to-end PDF extraction tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/venv/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
lint.select      = ["E", "F"]
lint.ignore      = []
fix         = true
